Conflicts:
	PA036-sprava_uziv_profilov/src/main/java/cz/muni/fi/pa036/nosql/persistence/RestaurantPersistence.java

+++ persistence/src/main/java/cz/muni/fi/pa036/dao/RestaurantPersistenceDao.java
@@ -29,7 +29,7 @@ public interface RestaurantPersistenceDao {
     Restaurant findByName(String name);
 
     
-    Restaurant findById(Long id);
+    Restaurant findById(int id);
     
     double getRating(int restaurantId);
